infrastructure/azure/requirements.txt
# Azure Infrastructure Requirements

# Core Azure SDK
azure-identity>=1.15.0
azure-core>=1.30.0
azure-mgmt-resource>=23.0.0

# Storage
azure-storage-blob>=12.19.0
azure-storage-queue>=12.6.0

# For testing
moto>=5.0.0

# Load dotenv

python-dotenv>=1.2.2

